extract base test class
