+++ core/src/test/java/com/spaceprogram/simplejpa/papeeria/BooleanObjectBuilderTest.java
@@ -1,38 +1,21 @@
 package com.spaceprogram.simplejpa.papeeria;
 
-import com.amazonaws.services.simpledb.AmazonSimpleDB;
-import com.amazonaws.services.simpledb.model.DeleteDomainRequest;
-import com.spaceprogram.simplejpa.EntityManagerFactoryImpl;
-import com.spaceprogram.simplejpa.EntityManagerSimpleJPA;
 import com.spaceprogram.simplejpa.papeeria.models.BooleanBox;
-import junit.framework.TestCase;
 
 import javax.persistence.EntityManager;
-import java.util.ArrayList;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
 
 /**
  * @author gkalabin@papeeria.com
  */
-public class BooleanObjectBuilderTest extends TestCase {
-  public static final String PERSISTENCE_UNIT_NAME = "papeeriatestunit";
-  private EntityManagerFactoryImpl myEntityManagerFactory;
-
-  private static final List<Class<?>> CLASSES = new ArrayList<Class<?>>();
-
+public class BooleanObjectBuilderTest extends BasePapeeriaTest {
   static {
-    CLASSES.add(BooleanBox.class);
+    ourTestClasses.add(BooleanBox.class);
   }
-
 
   @Override
   protected void setUp() throws Exception {
     super.setUp();
-    myEntityManagerFactory = new EntityManagerFactoryImpl(PERSISTENCE_UNIT_NAME, null, null, getStringClassNames());
-
-    EntityManager em = myEntityManagerFactory.createEntityManager();
+    EntityManager em = getEntityManager();
     {
       // create initial structure
       em.persist(new BooleanBox("t1", true));
@@ -41,30 +24,15 @@ public class BooleanObjectBuilderTest extends TestCase {
     }
   }
 
-  @Override
-  protected void tearDown() throws Exception {
-    super.tearDown();
-    EntityManagerSimpleJPA em = (EntityManagerSimpleJPA) myEntityManagerFactory.createEntityManager();
-    for (Class<?> aClass : CLASSES) {
-      AmazonSimpleDB db = em.getSimpleDb();
-      String domainName = em.getDomainName(aClass);
-
-      System.out.println("deleting domain: " + domainName);
-      DeleteDomainRequest deleteDomainRequest = new DeleteDomainRequest(domainName);
-      db.deleteDomain(deleteDomainRequest);
-    }
-    em.close();
-  }
-
   public void testSameObjects() {
     BooleanBox t1, anotherT1;
     {
-      EntityManager em = myEntityManagerFactory.createEntityManager();
+      EntityManager em = getEntityManager();
       t1 = em.find(BooleanBox.class, "t1");
       em.close();
     }
     {
-      EntityManager em = myEntityManagerFactory.createEntityManager();
+      EntityManager em = getEntityManager();
       anotherT1 = em.find(BooleanBox.class, "t1");
       em.close();
     }
@@ -76,25 +44,17 @@ public class BooleanObjectBuilderTest extends TestCase {
   public void testSameObjectFields() {
     BooleanBox t1, t2;
     {
-      EntityManager em = myEntityManagerFactory.createEntityManager();
+      EntityManager em = getEntityManager();
       t1 = em.find(BooleanBox.class, "t1");
       em.close();
     }
     {
-      EntityManager em = myEntityManagerFactory.createEntityManager();
+      EntityManager em = getEntityManager();
       t2 = em.find(BooleanBox.class, "t2");
       em.close();
     }
     assertTrue(t1.getBooleanObject());
     assertTrue(t2.getBooleanObject());
     assertTrue(t1.getBooleanObject() == t2.getBooleanObject());
-  }
-
-  private Set<String> getStringClassNames() {
-    Set<String> classNames = new HashSet<String>(CLASSES.size());
-    for (Class aClass : CLASSES) {
-      classNames.add(aClass.getName());
-    }
-    return classNames;
   }
 }
