+++ core/src/test/java/com/spaceprogram/simplejpa/papeeria/query/AbstractQueryTests.java
@@ -1,66 +1,34 @@
 package com.spaceprogram.simplejpa.papeeria.query;
 
-import com.amazonaws.services.simpledb.AmazonSimpleDB;
-import com.amazonaws.services.simpledb.model.DeleteDomainRequest;
-import com.spaceprogram.simplejpa.EntityManagerFactoryImpl;
-import com.spaceprogram.simplejpa.EntityManagerSimpleJPA;
+import com.spaceprogram.simplejpa.papeeria.BasePapeeriaTest;
 import com.spaceprogram.simplejpa.papeeria.models.PrimitiveBox;
-import junit.framework.TestCase;
 
 import javax.persistence.EntityManager;
 import javax.persistence.Query;
-import java.util.ArrayList;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
 
 /**
- * @author gkalabin@bardsoftware.com
+ * @author gkalabin@papeeria.com
  */
-public class AbstractQueryTests extends TestCase {
-  public static final String PERSISTENCE_UNIT_NAME = "papeeriatestunit";
-  private EntityManagerFactoryImpl myEntityManagerFactory;
-
-  private static final List<Class<?>> CLASSES = new ArrayList<Class<?>>();
-
+public class AbstractQueryTests extends BasePapeeriaTest {
   static {
-    CLASSES.add(PrimitiveBox.class);
+    ourTestClasses.add(PrimitiveBox.class);
   }
-
 
   @Override
   protected void setUp() throws Exception {
     super.setUp();
-    myEntityManagerFactory = new EntityManagerFactoryImpl(PERSISTENCE_UNIT_NAME, null, null, getStringClassNames());
-
-    EntityManager em = myEntityManagerFactory.createEntityManager();
-    {
-      // create initial structure
-      PrimitiveBox obj = new PrimitiveBox("foo", 42, 115L, .5d);
-      PrimitiveBox obj2 = new PrimitiveBox("bar", 24, 511L, 5.5d);
-      em.persist(obj);
-      em.persist(obj2);
-      em.close();
-    }
-  }
-
-  @Override
-  protected void tearDown() throws Exception {
-    super.tearDown();
-    EntityManagerSimpleJPA em = (EntityManagerSimpleJPA) myEntityManagerFactory.createEntityManager();
-    for (Class<?> aClass : CLASSES) {
-      AmazonSimpleDB db = em.getSimpleDb();
-      String domainName = em.getDomainName(aClass);
-
-      System.out.println("deleting domain: " + domainName);
-      DeleteDomainRequest deleteDomainRequest = new DeleteDomainRequest(domainName);
-      db.deleteDomain(deleteDomainRequest);
-    }
+    EntityManager em = getEntityManager();
+    // create initial structure
+    PrimitiveBox obj = new PrimitiveBox("foo", 42, 115L, .5d);
+    PrimitiveBox obj2 = new PrimitiveBox("bar", 24, 511L, 5.5d);
+    em.persist(obj);
+    em.persist(obj2);
     em.close();
   }
 
   public void testQueryWithIntParam() {
-    EntityManager em = myEntityManagerFactory.createEntityManager();
+    EntityManager em = getEntityManager();
     Query query = em.createQuery("SELECT b FROM PrimitiveBox b WHERE b.int=:int");
     query.setParameter("int", 42);
     List<PrimitiveBox> resultList = query.getResultList();
@@ -72,7 +40,7 @@ public class AbstractQueryTests extends TestCase {
   }
 
   public void testQueryWithLongParam() {
-    EntityManager em = myEntityManagerFactory.createEntityManager();
+    EntityManager em = getEntityManager();
     Query query = em.createQuery("SELECT b FROM PrimitiveBox b WHERE b.long<>:long");
     query.setParameter("long", 511L);
     List<PrimitiveBox> resultList = query.getResultList();
@@ -84,7 +52,7 @@ public class AbstractQueryTests extends TestCase {
   }
 
   public void testQueryWithDoubleParam() {
-    EntityManager em = myEntityManagerFactory.createEntityManager();
+    EntityManager em = getEntityManager();
     Query query = em.createQuery("SELECT b FROM PrimitiveBox b WHERE b.double=:double");
     query.setParameter("double", .5d);
     List<PrimitiveBox> resultList = query.getResultList();
@@ -93,13 +61,5 @@ public class AbstractQueryTests extends TestCase {
     assertEquals(115L, resultList.get(0).getLong());
     assertEquals(.5d, resultList.get(0).getDouble());
     em.close();
-  }
-
-  private Set<String> getStringClassNames() {
-    Set<String> classNames = new HashSet<String>(CLASSES.size());
-    for (Class aClass : CLASSES) {
-      classNames.add(aClass.getName());
-    }
-    return classNames;
   }
 }
